requirements.txt
pandas==1.4.4
setuptools
requests
pyyaml
deepl
unidecode
ckanapi
bs4
selenium
selenium-wire
deepl
ftfy